ci/ui: Add Cypress integration tests
We add these to the existing ui-unit-tests stage, thus installing
Cypress dependencies in the worker image.
Note that we need some complex management to handle starting the
application in the background and cleaning it up after Cypress tests are
complete (if we don't clean up, the Eve step hangs).

See: #2878