ci/build: Allocate even more RAM
We had issues with building the `metalk8s-ui` image in CI, likely due to
concurrent execution. The error was:

```
The build failed because the process exited too early. This probably
means the system ran out of memory or someone called `kill -9` on the
process.
```

To fix this, we add some memory to the Docker sidecar for the worker of
the `build` stage.