ci/ui: Run application from a sidecar for Cypress
Instead of running the metalk8s-ui application using the development
server (which watches changes to the source files and hot-reloads), we
instead build an NGINX image containing the statically built application
and deploy it as a sidecar to the Eve worker.
To better isolate this effort, we add a new stage for running UI
integration tests.

See: #2878