ui/cypress: Cancel requests after each test
We added an afterEach hook to call `window.gc`, though this doesn't fix
our issue with ongoing network requests being cancelled when new tests
start (which creates turmoil in the application).
We thus replace this hook with a more naive approach, which redirects
the browser to an `about:blank` page, to effectively cancel all ongoing
requests before a new test begins.

Note that this behaviour is preventing one from using the app after a
test case through Cypress runner.

See: #2878