salt/solutions: Never raise for absent config file
We used to raise a CommandExecutionError when calling
`metalk8s_solutions.read_config` when the file was absent and we didn't
pass `create=True`.
This does not make much sense: an absent config file simply means no
Solution configured.
We thus return the default (empty) configuration if the file is absent.

See: #2952