salt/solutions: Remove "available" from pillar
This list was obtained through a remote execution on the Bootstrap
minion, which made the ext_pillar sensible to degraded performance of
either Salt master or this minion, causing more issues down the line.
Since we only require this value when managing Solutions from the Salt
master or when templating some manifests on the Bootstrap minion, we
just remove this key and call the execution function directly instead.

See: #2952