eve: Make multi-node and single-node stage more generic
Multi-node stage could be use to do an expansions of various number of
node, so add the support from eve to generate "numbers" of node using
property as input.
Single-node and multi-node stage could be use to do installation with
different OS, so add the support from eve to install different OS using
property as input.