docs: fix `SPHINXOPTS` handling
The `SPHINXOPTS` value was not actually taken from the environment by
the `Makefile`, hence the value set in `tox.ini` wasn't really being
used at all.

Instead, hard-code some basic `SPHINXOPTS` in the `Makefile` (and
`make.bat`), ensure `O` is actually passed on through an environment
variable, and set a value we actually want to (only) set when running
the `docs` Tox target (`-W`, treating warnings as errors) using it.