buildchain, tox: rework cross-platform dependency handling
Instead of having both a `buildchain/requirements-Linux.txt` file and
`buildchain/requirements-Darwin.txt`, rendered from
`buildchain/requirements.in` and as such being almost identical (modulo
the platform-specific `pyinotify` and `macfsevents` modules), making
updates to the requirement lock files cumbersome (need to be executed on
all platforms), this commit creates a file containing a pinned version
of these two dependencies including a (manually added) platform
constraint, which is then used as a constraint file for `pip-compile` to
do its job (hence the `-c platform-requirements.txt` line in
`requirements.in`). After `pip-compile` generates a temporary lockfile
(cfr. code in `tox.ini`), we parse out the `pyinotify` and/or
`macfsevents` dependency, and append the constraint file to the result.

Using this approach we end up with a single
`buildchain/requirements.txt` which has all cross-platform `doit` (and
other) dependencies from `buildchain/requirements.in`, and
(platform-constrained) mentions of `pyinotify` and `macfsevents`.

The versions of the latter are pinned in `platform-requirements.txt`, so
these won't be upgraded when running `tox -e pip-compile`. If need be,
they can be bumped manually.

See: https://github.com/scality/metalk8s/pull/2986#issuecomment-748019832