salt: Increase timeout when waiting for containers
When creating `salt-master` and `repositories` static pods, we wait for
some containers to be in "Running" state before continuing. The default
timeout being set to 60s, this step may fail if the underlying platform
is too slow.
We raise this timeout to 180s (3 minutes) to avoid unnecessary failures
(if the container comes up earlier, the state `cri.wait_container` state
will exit as soon as possible).