ui: Add a flow typed client generator
The goal of this commit is to introduce a generator of a client based upon `@kubernetes/client` custom objects client from kubernetes `CustomResourceDefinition`.

The idea behind this is to have always up to date types definition for our custom resource definition and be in sync with the json schema contained within those definitions.

Refs: #2999