buildchain: embed ISO MD5 in the generated image
The `isomd5sum` tools allow to embed a hash of data sectors found in an
ISO file in an otherwise unused section, hence allowing the integrity of
(the data sectors of) an ISO image to be checked. This is, e.g., also
done for RHEL/CentOS/Fedora ISOs (their integrity can be checked at boot
time).

This commit adds a call to `implantisomd5` at the end of the buildchain,
right after the ISO is created and before its SHA256 is calculated.

Given this, one can run `checkisomd5` on a resulting ISO file after
download to ensure it's not corrupted.

Also mention the check in the docs, the `implantisomd5` requirement in
the developer docs, and add some code in the CI scripts to validate the
ISO using the embedded MD5 if present, next to the `SHA256SUM`.

Fixes: #3026
See: https://github.com/scality/metalk8s/issues/3026
See: https://github.com/rhinstaller/isomd5sum