dependencies: Configure dependabot to autoupdates core-ui version
We would like core-ui version to be automatically updated in all the projects that makes usse of it in order to benefit from latest fixes and feature.
Configuring Dependabot to do so would enable us to automatically get pull requests on ui projects as soon as a new version of core ui is available, easing communication of new version availabe accross UI chapter and enabling each squad to easily test their UI with the latest core-ui version.