buildchain,ui,deployment: use `nginx-alpine`
Instead of using the 'standard' `docker.io/nginx` version, use the
`-alpine` variant (and use a newer version).

Rationale is to reduce the size of the ISO we ship, since we're not
using any fancy `nginx` features in the first place, as well as space
consumed by images pulled by `containerd`.

Results (`du -s ...` of the directories):

```
$ cat before
1613124	_build/root/
1323720	_build/root/images/
43716	_build/root/images/nginx
110292	_build/root/images/nginx-1.15.8.tar
61460	_build/root/images/metalk8s-ui

$ cat after
1456000	_build/root/
1166600	_build/root/images/
9520	_build/root/images/nginx
23312	_build/root/images/nginx-1.19.6-alpine.tar
25516	_build/root/images/metalk8s-ui
```

Or, a total reduction of +- 154MB.