salt: Do not restart `salt-minion` process in background
Time to time when restarting the salt-minion process, restart take a bit
of time (more than 10s) so we have 2 salt command trying to start the
salt-minion process at the same time and sometime it happens that at the
end the salt-minion is not running at all `An instance is already running`.
Because the `salt-call --local service.restart salt-minion` first try to
restart the salt-minion process and then `service.running` seems to
happen when salt-minion process is actually not yet restarted.
To fix do not call the `service.restart` command in background so that
we wait for this to finish before running the `service.running` state.
NOTE: That this `bg: True` is not needed here since anyway in this SLS
we wait for the salt-minion process to restart