tests: Remove unneeded assertion for ssh_config
In a06bda1, we made the `ssh_config` argument to `kube_utils.get_pods`
optional, and added an assertion that it was never set to `None` if the
`node` argument was set (as it is used to compute the actual node name
from the SSH config file host).
However, the `utils.get_node_name` method already handles a `None`
ssh_config, so this assertion actually breaks behaviour when no
`ssh_config` is passed to pytest.

This reverts this change, which should fix tests run with `tox -e
tests-local`.