ui: Fix kubernetes build issue
When installing kubernetes client, the library is being built in the context of the consumer project meaning that it has access to the consumer node_modules.
That bring the issue that typescript was trying to build metalk8s types dependency to build kubernetes client leading to a build failure.

The issue has been resolved by this commit : scality/kubernetes-client-javascript@579d06628b6c15648551eab591e80b0bbd7990f3 which lists explicitly the types dependency of kubernetes client.