eve: Check status code when retrieving ISO images
In CI when we retrieve ISO images we do not fail properly if, for
example the ISO image does not exists (404 Not found).
So check the status code of the curl request.

NOTE: We do not use `-f/--fail` of curl as we do not want to retry if we
get a 404 for example