ui: Fix no banner on the deployed UI before prom volume provisioned
When we deploy a fresh bootstrap node, before provisioning the volume
for prometheus, we can't see the banner to guide you to provision the
volume from the deployed UI (not from local dev environment)

The root cause is how do we handle the prometheus error. The logic is
correct, we first check if there error from prometheus API, if no, we go
to check the deployment.

However, we can't tell the status of the HTTP request by just check if
`result.error.response` exist. It exists for all the out of 2XX
response. So we should check the response status instead of.