docs: Add context-dependent details to get support
This message is displayed in both "Installation / Troubleshooting" and
"Operation / Troubleshooting" chapters, so we factorize the content in a
single RST document, and include it where needed.

We also update the previously ill-named "scality-product" tag to
"scality_product", since tags should respect Python identifier rules.
We use this tag to include the relevant support details depending on the
context where this documentation will be served (either in
customer-facing documentation, or in open source like ReadTheDocs).