salt/tests: Add base `grains` to rendering context
We extract (and clean up) grains from a MetalK8s 2.7 bootstrap node,
running on CentOS 7.9.
These are stored in a YAML data source, which is read and exposed to the
rendering tests in a "base context" fixture.

Now, the only template which was not skipped, `top.sls.in`, can be
rendered without errors.
We can now expect that all non-skipped templates can be rendered with
their configured options, and update the test accordingly.

Running `tox -e unit-tests -- -m formulas` shows:
- 1 passed
- 235 skipped

See: #3073