salt/tests: Render reactor SLS with extra context
We define a new "extra_context" configuration option, which allows to
add arbitrary data to the rendering context.
This is especially useful for some specific formulas, like Reactors,
which expect a `data` value to contain the event for which they are
called.

Running `tox -e unit-tests -- -m formulas` shows:
- 3 passed
- 233 skipped

See: #3073