salt/tests: Check rendering for metalk8s.addons.nginx-ingress-control-plane.*
Add a "mode" option to control whether the formulas are rendered on a
minion or a master. For this option to have any effect, we add an `opts`
object to context, and customize it with this option value.

This allows us to cover all branches for
metalk8s.addons.nginx-ingress-control-plane.control-plane-ip.

Add a mock for saltutil.runner (which naively forwards calls to the
relevant mocked functions), and update the mine.get mock to allow
retrieving mine data for remote minions (no assumption on `tgt`).

Running `tox -e unit-tests -- -m formulas` shows:
- 184 passed
- 52 skipped

See: #3073