tests: Clean-up Dex-related tests
Mainly remove the `request_retry_session` fixture and replace it with
direct calls to `utils.request_retry_session` when needed, so one can
adjust the expected error codes and the number of retries depending on
the use-case.
Adding `502` and `504` to the list of expected error codes, and bumping
retries from 3 to 5, for all requests specifically targeted at Dex.