salt/etcd: Increase number of check attempts
When checking the cluster health after some live reconfiguration of the
etcd cluster, we were only attempting the check 5 times, with 10s of
sleep time between each. This totals to a 40s window during which things
need to stabilize, which, given the sensitivity of etcd, can be too
short.

We increase drastically the number of attempts to 31, so the maximum
period we will wait amounts to 5 minutes. This may be slightly high when
some real issue exists, but it gives us some headroom, and is much
better than failing because we're too optimistic.