salt: Add custom module metalk8s_network.routes
This module is a simplified version of
Salt network.routes, only handling IPv4 routes
for Linux dists using `ip` utility.

The goal of the module is to workaround a bug
in the original one when there is a blackhole
route defined
(see: https://github.com/saltstack/salt/issues/60126)
Once this bug is fixed upstream, we can remove it.

Refs: #3349