build: Calculate dependencies of LocalImage tasks
The previous approach consisted in:
- a static listing of these deps in the `file_dep` task attribute
- a dynamic verification from parsing the Dockerfile

This dynamic verification was however broken, because it used the
Dockerfile parent directory instead of the build context for listing the
dependencies.

When fixing this issue, we realized that a dynamic check couldn't work
for dependencies generated by other tasks, since the static `file_dep`
could not see them at task creation time.

To work around this limitation, and make the whole experience more
comfortable to developers, instead of checking, we use `doit`'s
`calc_dep` task attribute to dynamically compute dependencies using a
dedicated task, which can then wait for other tasks to complete.

See: #3271
See: #926