ui: Salt fix error handling when no job exists with provided id.
While working on the UI I met the following error :

```
TypeError: Cannot read property 'return' of undefined
    at getJobStatusFromPrintJob (utils.js:89)
    at refreshJobStatus$ (salt.js:188)
```

This is due to the fact that salt API was returning an empty object in `Result`.

Below  is the row result given by salt API.

```json
{
  "return": [
    {
      "20210802092226728983": {
        "Error": "Cannot contact returner or no job with this jid",
        "StartTime": "2021, Aug 02 09:22:26.728983",
        "Result": {}
      }
    }
  ]
}

This commit is checking wether the `Result` object is empty before trying to parse the error.

```