charts, salt: List and check dashboard UIDs
The dashboards we provision with MetalK8s have (or should have, which is
why some of this commit's changes are setting some UIDs) a UID. Starting
with Grafana v8.0, linking dashboards using their slug is not available
anymore (see
https://grafana.com/docs/grafana/latest/release-notes/release-notes-8-0-0/#breaking-changes
).

To make sure we can use these UIDs safely in links (mostly from MetalK8s
UI), we start by listing all of them in a JSON file, and enforce that
they are correct when rendering Helm charts.

Verifying the final deployed dashboards will be done in a subsequent
commit, through a post-install E2E test.

See: #3475