build: Simplify deployment of vendored dashboards
We previously stored dashboards in `charts/`, while they are not
retrieved from a Helm chart, simply vendored next to them.
It makes more sense to keep such dashboards vendored next to the
rendered `chart.sls`, as was already done for
`metalk8s.addons.prometheus-operator.deployed.dashboards`.
We factorize this approach and apply it to logging and ingress
dashboards.