tools/alert-tree: Define models for building trees
Using `ExistingAlert` and `DerivedAlert`, one can build a hierarchy of
alerts, with automatic definition of the PromQL expression required to
derive an alert from its children, and some other helpers for
manipulating such alert hierarchies.

Each one maps to an `AlertRule` serializable container, although
`ExistingAlert`s are not intended to be used like this, but only as a
simplified container of an alert name and its labels. This allows to
build complete PromQL queries when referencing such alerts in their
parent definitions.