lib-alert-tree: Add group by label filters to childrenJsonPath annotation
Some Parent alerts may be deduplicated for by components instances (e.g a VolumeAtRisk alert is raised for each pvc). To build a correct tree representation we then need to filter the child alerts matching those additional labels.
When defining an alert through a Prometheus rule, we can use a templating language in annotations to retrieve labels values of the expression that raised the alert.
In this commit we are using this templating language to make a reference to build a filter that only matches responsible children for a logical one.