salt: Configure Loki memberlist
We use memberlist so that each Loki instance
can talk with each other and replicate logs
received.
We ensure we have at least (n_nodes / 2) + 1
(rounded to the lower integer), so it ensures
data resiliency and allows to lose some instances
depending on the number of Loki replicas.

Re-render Loki charts to patch Loki StatefulSet
to add memberlist port:
```
    ./charts/render.py loki --namespace metalk8s-logging \
      --service-config loki metalk8s-loki-config \
      metalk8s/addons/logging/loki/config/loki.yaml metalk8s-logging  \
      --patch 'StatefulSet,metalk8s-logging,loki,spec:template:spec:containers:0:ports:1,
      {"name": "memberlist", "containerPort": 7946, "protocol": "TCP"}' \
      --remove-manifest Secret loki charts/loki.yaml charts/loki/ \
      > salt/metalk8s/addons/logging/loki/deployed/chart.sls
```