salt: Change replication Job names
We were using the nodeName in the Job name
but the nodeName can be very long, meaning
the Job name could exceed the limit of 63
characters, thus making it impossible to
run.
We now generate a random ID of 15 chars.
It also allows to keep the previous backup
Job a little longer (1 day), to be able to
retrieve logs from the previous replication
run without any risk of conflict if a new
one had to be scheduled.