charts,salt,tests,docs: Add ability to set replicas CP ingress controller
This commit add ability to configure the number of replicas for the
control plane ingress controller from bootstrap config.

NOTE: This is only used when MetalLB is enabled as otherwise the control
plane ingress controller is deployed as a DaemonSet.

NOTE: As a side effect of replicaCount change in the helm chart it's
remove the PodDisruptionBudget object, but this one was not used for the
moment

Render chart to salt state using
```
./charts/render.py ingress-nginx-control-plane --namespace metalk8s-ingress \
  charts/ingress-nginx-control-plane-deployment.yaml charts/ingress-nginx/ \
  > salt/metalk8s/addons/nginx-ingress-control-plane/deployed/chart-deployment.sls
```