salt,build: Mark the `pause` image as `pod infra container image`
We marks the `pause` image used by `containerd` as `pod infra container
image` for kubelet, so that kubelet does not remove this image.

We also put `pause` image in the embedded registry, so that we are sure
that node does not fetch this image from the internet.

NOTE: Since we fetch the pause image from the embedded registry we no
longer need to import it manually with salt on every node but only for
the registry node

Fixes: #3601