salt: Add a udevadm trigger when `persistentPath` does not exists
In some cases, a volume may be properly created but the path symlink (
`/dev/disks/by-XXX/`) does not exist because udevadm did not get
triggered on this device for... whatever reason.

To avoid that kind of issue we now trigger udevadm manually in the salt
execution module when the `persistentPath` does not exist and the
volume is actually "prepared".

Fixes: #3341