build/storage-operator: Copy previous manifests
The new layout suggests using kustomize and changes a lot the names of
generated objects, which we don't want to do for now.
Instead, we simply copy the old manifests over, with minor tweaks where
required (mainly the Deployment and an additional Role and RoleBinding
for leader election).
We also add labels to operator image.

See: #3517