build, storage-operator: Replace with new version
The migration has been completed, this commits replaces the previous
`storage-operator` directory with the migrated layout, and updates the
buildchain to account for these changes.
One thing to note is that we don't use a specific target for building
the operator image anymore, and simply rely on its generated (and
slightly updated) Dockerfile.

Closes: #3517