salt,tests: Allow to resolve registry endpoint from containers
In some cases, we may want to get some "stuff" from the internal registry
from inside a container (for example an operator that try to retrieve
some information from the registry).

Since we use a "fake" endpoint in containerd config file to reach the
registries, this commit make this "fake" endpoint also available when
resolving name with CoreDNS as we `rewrite` the "fake" endpoint name to
resolve to a non-headless service that points to the registry.

NOTE: We need to use a non-headless service as we want to forward the
port so that we have the same behavior as the fake endpoint configured
in containerd