charts,salt,tests: Make Ingress controller watch Ingress without class
The Workload Plane Ingress is the default Ingress controller that should
be used for Ingress object, so let's make this Ingress controller watch
Ingress object with no class.

NOTE: It shouldn't be needed since we define a default IngressClass
object, but during upgrade the Ingress object created without Class
in MetalK8s 2.10.x are ignored after upgrade with a message that the
Ingress object do not have any Class.

Re-render the state using:
```
./charts/render.py ingress-nginx --namespace metalk8s-ingress \
  charts/ingress-nginx.yaml charts/ingress-nginx/ \
  > salt/metalk8s/addons/nginx-ingress/deployed/chart.sls
```