scripts,salt: Allow ISO removal in iso-manager
This adds a `-r`/`--rm-archive` option to the script, and deprecates the
`--archive` in favor of `--add-archive` (for disambiguation).

It also slightly reworks our module extracting info from the
`product.txt` file in archives, to ensure the contents are valid at this
stage, instead of having error management in the
`metalk8s.archives.mounted` state.