build: Change base to rockylinux for metalk8s-utils
This will reduce the chances of being affected by CVEs on CentOS 7 not
being resolved in upstream repos, and ensure a more consistent
experience with the more common Red Hat 8 distributions which will be
used under MetalK8s.

Note that we also remove the `jnettop` tool (which is old, and not
available in EPEL 8), and install `iftop` instead.