salt/cri: Add utils to wait for kube-apiserver
When waiting for kube-apiserver to restart, we can't rely on K8s API to
give us a proper status (in single node, it would not be reachable).

However, we have observed that sometimes, changing its manifest wouldn't
trigger a restart by kubelet, which leads to a broken state where
nothing converges (again, in single node).

So we add simple utility methods to the `cri` module, and use it in
`metalk8s.kubernetes.apiserver.installed` to ensure we at least attempt
to restart kubelet if nothing moved after a while.