salt: Remove metalk8s.volumes from highstate
This was added to ensure loop devices backed by sparse files had an
updated logic after upgrading (with systemd units).

The side-effect is that upgrading a system where device paths have
changed (e.g. because of a reboot and different load order) can break,
since running `metalk8s_volumes.is_prepared` on such devices will return
False, and raise when we try to format them.

Since this is not needed anymore (the units will already exist), and can
be added back in the future if required, it is safe to remove for now.