salt,charts: Allow defining registry separately
Some charts now expect the image registry to be defined separately from
the repository, and enforces that these values are joined with a slash.
This causes issues with our macro "build_image_name", which builds the
whole path.
We add an option to this macro to omit the registry endpoint, and make
this value available to rendered charts using the charts/render.py
script header.