create useFederationsRoutes
The goal of this PR is to prepare the ability to load several versions of the same micro app
and be able to load micro app from a micro app.
To achieve that, we need to  :
- Be able to rely only on the discoveryUrl to render the Route instead of mixing it with the value of the config.json.
- Rework the way we render the micro app's `Routes`
 (see: `InternalRouter` in `shell-ui/src/FederatedApp.tsx`)

In the previous implementation we use `useDiscoveredViews` to render the Routes and render the navbar entries.
This was convenient because we reuse the hooks for the rendering the `Navbar` (see: `useFederatedNavbarEntries`)
and rendering the routes (DRY) but also cause some unnecessary coupling.
The route need to be exposed in the navbar in order to be accessible in the UI.

In a close futur, we want to be able to render some FederatedComponent without it to be exposed in the navbar.
For example, we might want to switch micro app with a selector instead of click on the navbar.

Here are some notable change made :
- we keep `useDiscoveredViews` for the navbar entries
- we create `useFederatedRoutes` to compute the federatedRoutes
- we add the url from where the micro app has been fetch in the metadata (cf: `EnrichedBuildtimeWebFinger`)
in order to be able to find the right micro app config since we can now have several time the same app.

Those changes simplify a lot the `FederatedApp.tsx` however we lost the capability to render only `Routes`
which the user have access. It means that we rely on the API to filter what the user can or not do (it's already the case).